Working through fix-73 and v0.6.0
