+++ index/spec_index.go
@@ -13,238 +13,44 @@
 package index
 
 import (
-	"errors"
 	"fmt"
-	"io/ioutil"
-	"net/http"
-	"strings"
-	"sync"
-
 	"github.com/pb33f/libopenapi/utils"
 	"github.com/vmware-labs/yaml-jsonpath/pkg/yamlpath"
 	"gopkg.in/yaml.v3"
+	"strings"
+	"sync"
 )
 
-// Constants used to determine if resolving is local, file based or remote file based.
-const (
-	LocalResolve = iota
-	HttpResolve
-	FileResolve
-)
-
-// Reference is a wrapper around *yaml.Node results to make things more manageable when performing
-// algorithms on data models. the *yaml.Node def is just a bit too low level for tracking state.
-type Reference struct {
-	Definition            string
-	Name                  string
-	Node                  *yaml.Node
-	ParentNode            *yaml.Node
-	Resolved              bool
-	Circular              bool
-	Seen                  bool
-	IsRemote              bool
-	RemoteLocation        string
-	Path                  string              // this won't always be available.
-	RequiredRefProperties map[string][]string // definition names (eg, #/definitions/One) to a list of required properties on this definition which reference that definition
-}
-
-// ReferenceMapped is a helper struct for mapped references put into sequence (we lose the key)
-type ReferenceMapped struct {
-	Reference  *Reference
-	Definition string
-}
-
-// SpecIndex is a complete pre-computed index of the entire specification. Numbers are pre-calculated and
-// quick direct access to paths, operations, tags are all available. No need to walk the entire node tree in rules,
-// everything is pre-walked if you need it.
-type SpecIndex struct {
-	allRefs                             map[string]*Reference                       // all (deduplicated) refs
-	rawSequencedRefs                    []*Reference                                // all raw references in sequence as they are scanned, not deduped.
-	linesWithRefs                       map[int]bool                                // lines that link to references.
-	allMappedRefs                       map[string]*Reference                       // these are the located mapped refs
-	allMappedRefsSequenced              []*ReferenceMapped                          // sequenced mapped refs
-	refsByLine                          map[string]map[int]bool                     // every reference and the lines it's referenced from
-	pathRefs                            map[string]map[string]*Reference            // all path references
-	paramOpRefs                         map[string]map[string]map[string]*Reference // params in operations.
-	paramCompRefs                       map[string]*Reference                       // params in components
-	paramAllRefs                        map[string]*Reference                       // combined components and ops
-	paramInlineDuplicates               map[string][]*Reference                     // inline params all with the same name
-	globalTagRefs                       map[string]*Reference                       // top level global tags
-	securitySchemeRefs                  map[string]*Reference                       // top level security schemes
-	requestBodiesRefs                   map[string]*Reference                       // top level request bodies
-	responsesRefs                       map[string]*Reference                       // top level responses
-	headersRefs                         map[string]*Reference                       // top level responses
-	examplesRefs                        map[string]*Reference                       // top level examples
-	securityRequirementRefs             map[string]map[string][]*Reference          // (NOT $ref) but a name based lookup for requirements
-	callbacksRefs                       map[string]map[string][]*Reference          // all links
-	linksRefs                           map[string]map[string][]*Reference          // all  callbacks
-	operationTagsRefs                   map[string]map[string][]*Reference          // tags found in operations
-	operationDescriptionRefs            map[string]map[string]*Reference            // descriptions in operations.
-	operationSummaryRefs                map[string]map[string]*Reference            // summaries in operations
-	callbackRefs                        map[string]*Reference                       // top level callback refs
-	serversRefs                         []*Reference                                // all top level server refs
-	rootServersNode                     *yaml.Node                                  // servers root node
-	opServersRefs                       map[string]map[string][]*Reference          // all operation level server overrides.
-	polymorphicRefs                     map[string]*Reference                       // every reference to a polymorphic ref
-	polymorphicAllOfRefs                []*Reference                                // every reference to 'allOf' references
-	polymorphicOneOfRefs                []*Reference                                // every reference to 'oneOf' references
-	polymorphicAnyOfRefs                []*Reference                                // every reference to 'anyOf' references
-	externalDocumentsRef                []*Reference                                // all external documents in spec
-	rootSecurity                        []*Reference                                // root security definitions.
-	rootSecurityNode                    *yaml.Node                                  // root security node.
-	refsWithSiblings                    map[string]Reference                        // references with sibling elements next to them
-	pathRefsLock                        sync.Mutex                                  // create lock for all refs maps, we want to build data as fast as we can
-	externalDocumentsCount              int                                         // number of externalDocument nodes found
-	operationTagsCount                  int                                         // number of unique tags in operations
-	globalTagsCount                     int                                         // number of global tags defined
-	totalTagsCount                      int                                         // number unique tags in spec
-	securitySchemesCount                int                                         // security schemes
-	globalRequestBodiesCount            int                                         // component request bodies
-	globalResponsesCount                int                                         // component responses
-	globalHeadersCount                  int                                         // component headers
-	globalExamplesCount                 int                                         // component examples
-	globalLinksCount                    int                                         // component links
-	globalCallbacksCount                int                                         // component callbacks
-	globalCallbacks                     int                                         // component callbacks.
-	pathCount                           int                                         // number of paths
-	operationCount                      int                                         // number of operations
-	operationParamCount                 int                                         // number of params defined in operations
-	componentParamCount                 int                                         // number of params defined in components
-	componentsInlineParamUniqueCount    int                                         // number of inline params with unique names
-	componentsInlineParamDuplicateCount int                                         // number of inline params with duplicate names
-	schemaCount                         int                                         // number of schemas
-	refCount                            int                                         // total ref count
-	root                                *yaml.Node                                  // the root document
-	pathsNode                           *yaml.Node                                  // paths node
-	tagsNode                            *yaml.Node                                  // tags node
-	componentsNode                      *yaml.Node                                  // components node
-	parametersNode                      *yaml.Node                                  // components/parameters node
-	allParametersNode                   map[string]*Reference                       // all parameters node
-	allParameters                       map[string]*Reference                       // all parameters (components/defs)
-	schemasNode                         *yaml.Node                                  // components/schemas node
-	allInlineSchemaDefinitions          []*Reference                                // all schemas found in document outside of components (openapi) or definitions (swagger).
-	allInlineSchemaObjectDefinitions    []*Reference                                // all schemas that are objects found in document outside of components (openapi) or definitions (swagger).
-	allComponentSchemaDefinitions       map[string]*Reference                       // all schemas found in components (openapi) or definitions (swagger).
-	securitySchemesNode                 *yaml.Node                                  // components/securitySchemes node
-	allSecuritySchemes                  map[string]*Reference                       // all security schemes / definitions.
-	requestBodiesNode                   *yaml.Node                                  // components/requestBodies node
-	allRequestBodies                    map[string]*Reference                       // all request bodies
-	responsesNode                       *yaml.Node                                  // components/responses node
-	allResponses                        map[string]*Reference                       // all responses
-	headersNode                         *yaml.Node                                  // components/headers node
-	allHeaders                          map[string]*Reference                       // all headers
-	examplesNode                        *yaml.Node                                  // components/examples node
-	allExamples                         map[string]*Reference                       // all components examples
-	linksNode                           *yaml.Node                                  // components/links node
-	allLinks                            map[string]*Reference                       // all links
-	callbacksNode                       *yaml.Node                                  // components/callbacks node
-	allCallbacks                        map[string]*Reference                       // all components examples
-	externalDocumentsNode               *yaml.Node                                  // external documents node
-	allExternalDocuments                map[string]*Reference                       // all external documents
-	externalSpecIndex                   map[string]*SpecIndex                       // create a primary index of all external specs and componentIds
-	refErrors                           []*IndexingError                            // errors when indexing references
-	operationParamErrors                []*IndexingError                            // errors when indexing parameters
-	allDescriptions                     []*DescriptionReference                     // every single description found in the spec.
-	allSummaries                        []*DescriptionReference                     // every single summary found in the spec.
-	allEnums                            []*EnumReference                            // every single enum found in the spec.
-	allObjectsWithProperties            []*ObjectReference                          // every single object with properties found in the spec.
-	enumCount                           int
-	descriptionCount                    int
-	summaryCount                        int
-	seenRemoteSources                   map[string]*yaml.Node
-	remoteLock                          sync.Mutex
-	circularReferences                  []*CircularReferenceResult // only available when the resolver has been used.
-	allowCircularReferences             bool                       // decide if you want to error out, or allow circular references, default is false.
-}
-
-// ExternalLookupFunction is for lookup functions that take a JSONSchema reference and tries to find that node in the
-// URI based document. Decides if the reference is local, remote or in a file.
-type ExternalLookupFunction func(id string) (foundNode *yaml.Node, rootNode *yaml.Node, lookupError error)
-
-// IndexingError holds data about something that went wrong during indexing.
-type IndexingError struct {
-	Error error
-	Node  *yaml.Node
-	Path  string
-}
-
-// DescriptionReference holds data about a description that was found and where it was found.
-type DescriptionReference struct {
-	Content   string
-	Path      string
-	Node      *yaml.Node
-	IsSummary bool
-}
-
-type EnumReference struct {
-	Node       *yaml.Node
-	Type       *yaml.Node
-	Path       string
-	SchemaNode *yaml.Node
-	ParentNode *yaml.Node
-}
-
-type ObjectReference struct {
-	Node       *yaml.Node
-	Path       string
-	ParentNode *yaml.Node
-}
-
-var methodTypes = []string{"get", "post", "put", "patch", "options", "head", "delete"}
-
-func runIndexFunction(funcs []func() int, wg *sync.WaitGroup) {
-	for _, cFunc := range funcs {
-		go func(wg *sync.WaitGroup, cf func() int) {
-			cf()
-			wg.Done()
-		}(wg, cFunc)
-	}
+// NewSpecIndexWithConfig will create a new index of an OpenAPI or Swagger spec. It uses the same logic as NewSpecIndex
+// except it sets a base URL for resolving relative references, except it also allows for granular control over
+// how the index is set up.
+func NewSpecIndexWithConfig(rootNode *yaml.Node, config *SpecIndexConfig) *SpecIndex {
+	index := new(SpecIndex)
+	index.config = config
+	boostrapIndexCollections(rootNode, index)
+	return createNewIndex(rootNode, index)
 }
 
 // NewSpecIndex will create a new index of an OpenAPI or Swagger spec. It's not resolved or converted into anything
 // other than a raw index of every node for every content type in the specification. This process runs as fast as
 // possible so dependencies looking through the tree, don't need to walk the entire thing over, and over.
+//
+// Deprecated: Use NewSpecIndexWithConfig instead, this function will be removed in the future because it
+// defaults to allowing remote references and file references. This is a potential security risk and should be controlled by
+// providing a SpecIndexConfig that explicitly sets the AllowRemoteLookup and AllowFileLookup to true.
+// This function also does not support specifications with relative references that may not exist locally.
+//  - https://github.com/pb33f/libopenapi/issues/73
 func NewSpecIndex(rootNode *yaml.Node) *SpecIndex {
 	index := new(SpecIndex)
-	index.root = rootNode
-	index.allRefs = make(map[string]*Reference)
-	index.allMappedRefs = make(map[string]*Reference)
-	index.refsByLine = make(map[string]map[int]bool)
-	index.linesWithRefs = make(map[int]bool)
-	index.pathRefs = make(map[string]map[string]*Reference)
-	index.paramOpRefs = make(map[string]map[string]map[string]*Reference)
-	index.operationTagsRefs = make(map[string]map[string][]*Reference)
-	index.operationDescriptionRefs = make(map[string]map[string]*Reference)
-	index.operationSummaryRefs = make(map[string]map[string]*Reference)
-	index.paramCompRefs = make(map[string]*Reference)
-	index.paramAllRefs = make(map[string]*Reference)
-	index.paramInlineDuplicates = make(map[string][]*Reference)
-	index.globalTagRefs = make(map[string]*Reference)
-	index.securitySchemeRefs = make(map[string]*Reference)
-	index.requestBodiesRefs = make(map[string]*Reference)
-	index.responsesRefs = make(map[string]*Reference)
-	index.headersRefs = make(map[string]*Reference)
-	index.examplesRefs = make(map[string]*Reference)
-	index.callbacksRefs = make(map[string]map[string][]*Reference)
-	index.linksRefs = make(map[string]map[string][]*Reference)
-	index.callbackRefs = make(map[string]*Reference)
-	index.externalSpecIndex = make(map[string]*SpecIndex)
-	index.allComponentSchemaDefinitions = make(map[string]*Reference)
-	index.allParameters = make(map[string]*Reference)
-	index.allSecuritySchemes = make(map[string]*Reference)
-	index.allRequestBodies = make(map[string]*Reference)
-	index.allResponses = make(map[string]*Reference)
-	index.allHeaders = make(map[string]*Reference)
-	index.allExamples = make(map[string]*Reference)
-	index.allLinks = make(map[string]*Reference)
-	index.allCallbacks = make(map[string]*Reference)
-	index.allExternalDocuments = make(map[string]*Reference)
-	index.securityRequirementRefs = make(map[string]map[string][]*Reference)
-	index.polymorphicRefs = make(map[string]*Reference)
-	index.refsWithSiblings = make(map[string]Reference)
-	index.seenRemoteSources = make(map[string]*yaml.Node)
-	index.opServersRefs = make(map[string]map[string][]*Reference)
+	index.config = &SpecIndexConfig{
+		AllowRemoteLookup: true,
+		AllowFileLookup:   true,
+	}
+	boostrapIndexCollections(rootNode, index)
+	return createNewIndex(rootNode, index)
+}
 
+func createNewIndex(rootNode *yaml.Node, index *SpecIndex) *SpecIndex {
 	// there is no node! return an empty index.
 	if rootNode == nil {
 		return index
@@ -253,9 +59,6 @@ func NewSpecIndex(rootNode *yaml.Node) *SpecIndex {
 	// boot index.
 	results := index.ExtractRefs(index.root.Content[0], index.root, []string{}, 0, false, "")
 
-	// pull out references
-	index.ExtractComponentsFromRefs(results)
-
 	// map poly refs
 	poly := make([]*Reference, len(index.polymorphicRefs))
 	z := 0
@@ -263,6 +66,9 @@ func NewSpecIndex(rootNode *yaml.Node) *SpecIndex {
 		poly[z] = index.polymorphicRefs[i]
 		z++
 	}
+
+	// pull out references
+	index.ExtractComponentsFromRefs(results)
 	index.ExtractComponentsFromRefs(poly)
 
 	index.ExtractExternalDocuments(index.root)
@@ -526,12 +332,12 @@ func (index *SpecIndex) GetParametersNode() *yaml.Node {
 }
 
 // GetReferenceIndexErrors will return any errors that occurred when indexing references
-func (index *SpecIndex) GetReferenceIndexErrors() []*IndexingError {
+func (index *SpecIndex) GetReferenceIndexErrors() []error {
 	return index.refErrors
 }
 
 // GetOperationParametersIndexErrors any errors that occurred when indexing operation parameters
-func (index *SpecIndex) GetOperationParametersIndexErrors() []*IndexingError {
+func (index *SpecIndex) GetOperationParametersIndexErrors() []error {
 	return index.operationParamErrors
 }
 
@@ -608,313 +414,7 @@ func (index *SpecIndex) checkPolymorphicNode(name string) (bool, string) {
 	return false, ""
 }
 
-// ExtractRefs will return a deduplicated slice of references for every unique ref found in the document.
-// The total number of refs, will generally be much higher, you can extract those from GetRawReferenceCount()
-func (index *SpecIndex) ExtractRefs(node, parent *yaml.Node, seenPath []string, level int, poly bool, pName string) []*Reference {
-	if node == nil {
-		return nil
-	}
-	var found []*Reference
-	if len(node.Content) > 0 {
-		var prev, polyName string
-		for i, n := range node.Content {
 
-			if utils.IsNodeMap(n) || utils.IsNodeArray(n) {
-				level++
-				// check if we're using  polymorphic values. These tend to create rabbit warrens of circular
-				// references if every single link is followed. We don't resolve polymorphic values.
-				isPoly, _ := index.checkPolymorphicNode(prev)
-				polyName = pName
-				if isPoly {
-					poly = true
-					if prev != "" {
-						polyName = prev
-					}
-				}
-				found = append(found, index.ExtractRefs(n, node, seenPath, level, poly, polyName)...)
-			}
-
-			// check if we're dealing with an inline schema definition, that isn't part of an array
-			// (which means it's being used as a value in an array, and it's not a label)
-			// https://github.com/pb33f/libopenapi/issues/76
-			if i%2 == 0 && n.Value == "schema" && !utils.IsNodeArray(node) && (i+1 < len(node.Content)) {
-				isRef, _, _ := utils.IsNodeRefValue(node.Content[i+1])
-				if isRef {
-					continue
-				}
-				ref := &Reference{
-					Node: node.Content[i+1],
-					Path: fmt.Sprintf("$.%s", strings.Join(seenPath, ".")),
-				}
-				index.allInlineSchemaDefinitions = append(index.allInlineSchemaDefinitions, ref)
-
-				// check if the schema is an object or an array,
-				// and if so, add it to the list of inline schema object definitions.
-				k, v := utils.FindKeyNodeTop("type", node.Content[i+1].Content)
-				if k != nil && v != nil {
-					if v.Value == "object" || v.Value == "array" {
-						index.allInlineSchemaObjectDefinitions = append(index.allInlineSchemaObjectDefinitions, ref)
-					}
-				}
-			}
-
-			if i%2 == 0 && n.Value == "$ref" {
-
-				// only look at scalar values, not maps (looking at you k8s)
-				if !utils.IsNodeStringValue(node.Content[i+1]) {
-					continue
-				}
-
-				index.linesWithRefs[n.Line] = true
-
-				fp := make([]string, len(seenPath))
-				for x, foundPathNode := range seenPath {
-					fp[x] = foundPathNode
-				}
-
-				value := node.Content[i+1].Value
-
-				segs := strings.Split(value, "/")
-				name := segs[len(segs)-1]
-				_, p := utils.ConvertComponentIdIntoFriendlyPathSearch(value)
-				ref := &Reference{
-					Definition: value,
-					Name:       name,
-					Node:       node,
-					Path:       p,
-				}
-
-				// add to raw sequenced refs
-				index.rawSequencedRefs = append(index.rawSequencedRefs, ref)
-
-				// add ref by line number
-				refNameIndex := strings.LastIndex(value, "/")
-				refName := value[refNameIndex+1:]
-				if len(index.refsByLine[refName]) > 0 {
-					index.refsByLine[refName][n.Line] = true
-				} else {
-					v := make(map[int]bool)
-					v[n.Line] = true
-					index.refsByLine[refName] = v
-				}
-
-				// if this ref value has any siblings (node.Content is larger than two elements)
-				// then add to refs with siblings
-				if len(node.Content) > 2 {
-					copiedNode := *node
-					copied := Reference{
-						Definition: ref.Definition,
-						Name:       ref.Name,
-						Node:       &copiedNode,
-						Path:       p,
-					}
-					// protect this data using a copy, prevent the resolver from destroying things.
-					index.refsWithSiblings[value] = copied
-				}
-
-				// if this is a polymorphic reference, we're going to leave it out
-				// allRefs. We don't ever want these resolved, so instead of polluting
-				// the timeline, we will keep each poly ref in its own collection for later
-				// analysis.
-				if poly {
-					index.polymorphicRefs[value] = ref
-
-					// index each type
-					switch pName {
-					case "anyOf":
-						index.polymorphicAnyOfRefs = append(index.polymorphicAnyOfRefs, ref)
-					case "allOf":
-						index.polymorphicAllOfRefs = append(index.polymorphicAllOfRefs, ref)
-					case "oneOf":
-						index.polymorphicOneOfRefs = append(index.polymorphicOneOfRefs, ref)
-					}
-					continue
-				}
-
-				// check if this is a dupe, if so, skip it, we don't care now.
-				if index.allRefs[value] != nil { // seen before, skip.
-					continue
-				}
-
-				if value == "" {
-
-					completedPath := fmt.Sprintf("$.%s", strings.Join(fp, "."))
-
-					indexError := &IndexingError{
-						Error: errors.New("schema reference is empty and cannot be processed"),
-						Node:  node.Content[i+1],
-						Path:  completedPath,
-					}
-
-					index.refErrors = append(index.refErrors, indexError)
-
-					continue
-				}
-
-				index.allRefs[value] = ref
-				found = append(found, ref)
-			}
-
-			if i%2 == 0 && n.Value != "$ref" && n.Value != "" {
-
-				nodePath := fmt.Sprintf("$.%s", strings.Join(seenPath, "."))
-
-				// capture descriptions and summaries
-				if n.Value == "description" {
-
-					// if the parent is a sequence, ignore.
-					if utils.IsNodeArray(node) {
-						continue
-					}
-
-					ref := &DescriptionReference{
-						Content:   node.Content[i+1].Value,
-						Path:      nodePath,
-						Node:      node.Content[i+1],
-						IsSummary: false,
-					}
-
-					index.allDescriptions = append(index.allDescriptions, ref)
-					index.descriptionCount++
-				}
-
-				if n.Value == "summary" {
-
-					var b *yaml.Node
-					if len(node.Content) == i+1 {
-						b = node.Content[i]
-					} else {
-						b = node.Content[i+1]
-					}
-					ref := &DescriptionReference{
-						Content:   b.Value,
-						Path:      nodePath,
-						Node:      b,
-						IsSummary: true,
-					}
-
-					index.allSummaries = append(index.allSummaries, ref)
-					index.summaryCount++
-				}
-
-				// capture security requirement references (these are not traditional references, but they
-				// are used as a look-up. This is the only exception to the design.
-				if n.Value == "security" {
-					var b *yaml.Node
-					if len(node.Content) == i+1 {
-						b = node.Content[i]
-					} else {
-						b = node.Content[i+1]
-					}
-					if utils.IsNodeArray(b) {
-						var secKey string
-						for k := range b.Content {
-							if utils.IsNodeMap(b.Content[k]) {
-								for g := range b.Content[k].Content {
-									if g%2 == 0 {
-										secKey = b.Content[k].Content[g].Value
-										continue
-									}
-									if utils.IsNodeArray(b.Content[k].Content[g]) {
-										var refMap map[string][]*Reference
-										if index.securityRequirementRefs[secKey] == nil {
-											index.securityRequirementRefs[secKey] = make(map[string][]*Reference)
-											refMap = index.securityRequirementRefs[secKey]
-										} else {
-											refMap = index.securityRequirementRefs[secKey]
-										}
-										for r := range b.Content[k].Content[g].Content {
-											var refs []*Reference
-											if refMap[b.Content[k].Content[g].Content[r].Value] != nil {
-												refs = refMap[b.Content[k].Content[g].Content[r].Value]
-											}
-
-											refs = append(refs, &Reference{
-												Definition: b.Content[k].Content[g].Content[r].Value,
-												Path:       fmt.Sprintf("%s.security[%d].%s[%d]", nodePath, k, secKey, r),
-												Node:       b.Content[k].Content[g].Content[r],
-											})
-
-											index.securityRequirementRefs[secKey][b.Content[k].Content[g].Content[r].Value] = refs
-										}
-									}
-								}
-							}
-						}
-					}
-				}
-				// capture enums
-				if n.Value == "enum" {
-
-					// all enums need to have a type, extract the type from the node where the enum was found.
-					_, enumKeyValueNode := utils.FindKeyNodeTop("type", node.Content)
-
-					if enumKeyValueNode != nil {
-						ref := &EnumReference{
-							Path:       nodePath,
-							Node:       node.Content[i+1],
-							Type:       enumKeyValueNode,
-							SchemaNode: node,
-							ParentNode: parent,
-						}
-
-						index.allEnums = append(index.allEnums, ref)
-						index.enumCount++
-					}
-				}
-				// capture all objects with properties
-				if n.Value == "properties" {
-					_, typeKeyValueNode := utils.FindKeyNodeTop("type", node.Content)
-
-					if typeKeyValueNode != nil {
-						isObject := false
-
-						if typeKeyValueNode.Value == "object" {
-							isObject = true
-						}
-
-						for _, v := range typeKeyValueNode.Content {
-							if v.Value == "object" {
-								isObject = true
-							}
-						}
-
-						if isObject {
-							index.allObjectsWithProperties = append(index.allObjectsWithProperties, &ObjectReference{
-								Path:       nodePath,
-								Node:       node,
-								ParentNode: parent,
-							})
-						}
-					}
-				}
-
-				seenPath = append(seenPath, n.Value)
-				prev = n.Value
-			}
-
-			// if next node is map, don't add segment.
-			if i < len(node.Content)-1 {
-				next := node.Content[i+1]
-
-				if i%2 != 0 && next != nil && !utils.IsNodeArray(next) && !utils.IsNodeMap(next) {
-					seenPath = seenPath[:len(seenPath)-1]
-				}
-			}
-		}
-		if len(seenPath) > 0 {
-			seenPath = seenPath[:len(seenPath)-1]
-		}
-
-	}
-	if len(seenPath) > 0 {
-		seenPath = seenPath[:len(seenPath)-1]
-	}
-
-	index.refCount = len(index.allRefs)
-
-	return found
-}
 
 // GetPathCount will return the number of paths found in the spec
 func (index *SpecIndex) GetPathCount() int {
@@ -1623,108 +1123,6 @@ func (index *SpecIndex) GetInlineUniqueParamCount() int {
 	return index.countUniqueInlineDuplicates()
 }
 
-// ExtractComponentsFromRefs returns located components from references. The returned nodes from here
-// can be used for resolving as they contain the actual object properties.
-func (index *SpecIndex) ExtractComponentsFromRefs(refs []*Reference) []*Reference {
-	var found []*Reference
-	for _, ref := range refs {
-
-		// check reference for backslashes (hah yeah seen this too!)
-		if strings.Contains(ref.Definition, "\\") { // this was from blazemeter.com haha!
-			_, path := utils.ConvertComponentIdIntoFriendlyPathSearch(ref.Definition)
-			indexError := &IndexingError{
-				Error: fmt.Errorf("component '%s' contains a backslash '\\'. It's not valid", ref.Definition),
-				Node:  ref.Node,
-				Path:  path,
-			}
-			index.refErrors = append(index.refErrors, indexError)
-			continue
-		}
-
-		located := index.FindComponent(ref.Definition, ref.Node)
-		if located != nil {
-			if index.allMappedRefs[ref.Definition] == nil {
-				found = append(found, located)
-				index.allMappedRefs[ref.Definition] = located
-				index.allMappedRefsSequenced = append(index.allMappedRefsSequenced, &ReferenceMapped{
-					Reference:  located,
-					Definition: ref.Definition,
-				})
-			}
-		} else {
-
-			_, path := utils.ConvertComponentIdIntoFriendlyPathSearch(ref.Definition)
-			indexError := &IndexingError{
-				Error: fmt.Errorf("component '%s' does not exist in the specification", ref.Definition),
-				Node:  ref.Node,
-				Path:  path,
-			}
-			index.refErrors = append(index.refErrors, indexError)
-		}
-	}
-	return found
-}
-
-// FindComponent will locate a component by its reference, returns nil if nothing is found.
-// This method will recurse through remote, local and file references. For each new external reference
-// a new index will be created. These indexes can then be traversed recursively.
-func (index *SpecIndex) FindComponent(componentId string, parent *yaml.Node) *Reference {
-	if index.root == nil {
-		return nil
-	}
-
-	// FIXME: This is a potential security hole, and needs to be made optional (see log4j fiasco)
-	remoteLookup := func(id string) (*yaml.Node, *yaml.Node, error) {
-		return index.lookupRemoteReference(id)
-	}
-
-	// FIXME: As above
-	fileLookup := func(id string) (*yaml.Node, *yaml.Node, error) {
-		return index.lookupFileReference(id)
-	}
-
-	switch DetermineReferenceResolveType(componentId) {
-	case LocalResolve: // ideally, every single ref in every single spec is local. however, this is not the case.
-		return index.FindComponentInRoot(componentId)
-
-	case HttpResolve:
-		uri := strings.Split(componentId, "#")
-		if len(uri) >= 2 {
-			return index.performExternalLookup(uri, componentId, remoteLookup, parent)
-		}
-		if len(uri) == 1 {
-			// if there is no reference, second segment is empty / has no name
-			// this means there is no component to look-up and the entire file should be pulled in.
-			// to stop all the other code from breaking (that is expecting a component), let's just post-pend
-			// a hash to the end of the componentId and ensure the uri slice is as expected.
-			// described in https://github.com/pb33f/libopenapi/issues/37
-			componentId = fmt.Sprintf("%s#", componentId)
-			uri = append(uri, "")
-			return index.performExternalLookup(uri, componentId, remoteLookup, parent)
-		}
-
-	case FileResolve:
-		uri := strings.Split(componentId, "#")
-		if len(uri) == 2 {
-			return index.performExternalLookup(uri, componentId, fileLookup, parent)
-		}
-		if len(uri) == 1 {
-			// if there is no reference, second segment is empty / has no name
-			// this means there is no component to look-up and the entire file should be pulled in.
-			// to stop all the other code from breaking (that is expecting a component), let's just post-pend
-			// a hash to the end of the componentId and ensure the uri slice is as expected.
-			// described in https://github.com/pb33f/libopenapi/issues/37
-			//
-			// ^^ this same issue was re-reported in file based lookups in vacuum.
-			// more info here: https://github.com/daveshanley/vacuum/issues/225
-			componentId = fmt.Sprintf("%s#", componentId)
-			uri = append(uri, "")
-			return index.performExternalLookup(uri, componentId, fileLookup, parent)
-		}
-	}
-	return nil
-}
-
 // GetAllDescriptionsCount will collect together every single description found in the document
 func (index *SpecIndex) GetAllDescriptionsCount() int {
 	return len(index.allDescriptions)
@@ -1733,557 +1131,4 @@ func (index *SpecIndex) GetAllDescriptionsCount() int {
 // GetAllSummariesCount will collect together every single summary found in the document
 func (index *SpecIndex) GetAllSummariesCount() int {
 	return len(index.allSummaries)
-}
-
-func DetermineReferenceResolveType(ref string) int {
-	if ref != "" && ref[0] == '#' {
-		return LocalResolve
-	}
-	if ref != "" && len(ref) >= 5 && (ref[:5] == "https" || ref[:5] == "http:") {
-		return HttpResolve
-	}
-	if strings.Contains(ref, ".json") ||
-		strings.Contains(ref, ".yaml") ||
-		strings.Contains(ref, ".yml") {
-		return FileResolve
-	}
-	return -1
-}
-
-/* private */
-
-func (index *SpecIndex) extractDefinitionsAndSchemas(schemasNode *yaml.Node, pathPrefix string) {
-	var name string
-	for i, schema := range schemasNode.Content {
-		if i%2 == 0 {
-			name = schema.Value
-			continue
-		}
-
-		def := fmt.Sprintf("%s%s", pathPrefix, name)
-		ref := &Reference{
-			Definition:            def,
-			Name:                  name,
-			Node:                  schema,
-			Path:                  fmt.Sprintf("$.components.schemas.%s", name),
-			ParentNode:            schemasNode,
-			RequiredRefProperties: index.extractDefinitionRequiredRefProperties(schemasNode, map[string][]string{}),
-		}
-		index.allComponentSchemaDefinitions[def] = ref
-	}
-}
-
-// extractDefinitionRequiredRefProperties goes through the direct properties of a schema and extracts the map of required definitions from within it
-func (index *SpecIndex) extractDefinitionRequiredRefProperties(schemaNode *yaml.Node, reqRefProps map[string][]string) map[string][]string {
-	if schemaNode == nil {
-		return reqRefProps
-	}
-
-	// If the node we're looking at is a direct ref to another model without any properties, mark it as required, but still continue to look for required properties
-	isRef, _, defPath := utils.IsNodeRefValue(schemaNode)
-	if isRef {
-		if _, ok := reqRefProps[defPath]; !ok {
-			reqRefProps[defPath] = []string{}
-		}
-	}
-
-	// Check for a required parameters list, and return if none exists, as any properties will be optional
-	_, requiredSeqNode := utils.FindKeyNodeTop("required", schemaNode.Content)
-	if requiredSeqNode == nil {
-		return reqRefProps
-	}
-
-	_, propertiesMapNode := utils.FindKeyNodeTop("properties", schemaNode.Content)
-	if propertiesMapNode == nil {
-		// TODO: Log a warning on the resolver, because if you have required properties, but no actual properties, something is wrong
-		return reqRefProps
-	}
-
-	name := ""
-	for i, param := range propertiesMapNode.Content {
-		if i%2 == 0 {
-			name = param.Value
-			continue
-		}
-
-		// Check to see if the current property is directly embedded within the current schema, and handle its properties if so
-		_, paramPropertiesMapNode := utils.FindKeyNodeTop("properties", param.Content)
-		if paramPropertiesMapNode != nil {
-			reqRefProps = index.extractDefinitionRequiredRefProperties(param, reqRefProps)
-		}
-
-		// Check to see if the current property is polymorphic, and dive into that model if so
-		for _, key := range []string{"allOf", "oneOf", "anyOf"} {
-			_, ofNode := utils.FindKeyNodeTop(key, param.Content)
-			if ofNode != nil {
-				for _, ofNodeItem := range ofNode.Content {
-					reqRefProps = index.extractRequiredReferenceProperties(ofNodeItem, name, reqRefProps)
-				}
-			}
-		}
-	}
-
-	// Run through each of the required properties and extract _their_ required references
-	for _, requiredPropertyNode := range requiredSeqNode.Content {
-		_, requiredPropDefNode := utils.FindKeyNodeTop(requiredPropertyNode.Value, propertiesMapNode.Content)
-		if requiredPropDefNode == nil {
-			continue
-		}
-
-		reqRefProps = index.extractRequiredReferenceProperties(requiredPropDefNode, requiredPropertyNode.Value, reqRefProps)
-	}
-
-	return reqRefProps
-}
-
-// extractRequiredReferenceProperties returns a map of definition names to the property or properties which reference it within a node
-func (index *SpecIndex) extractRequiredReferenceProperties(requiredPropDefNode *yaml.Node, propName string, reqRefProps map[string][]string) map[string][]string {
-	isRef, _, defPath := utils.IsNodeRefValue(requiredPropDefNode)
-	if !isRef {
-		_, defItems := utils.FindKeyNodeTop("items", requiredPropDefNode.Content)
-		if defItems != nil {
-			isRef, _, defPath = utils.IsNodeRefValue(defItems)
-		}
-	}
-
-	if /* still */ !isRef {
-		return reqRefProps
-	}
-
-	if _, ok := reqRefProps[defPath]; !ok {
-		reqRefProps[defPath] = []string{}
-	}
-	reqRefProps[defPath] = append(reqRefProps[defPath], propName)
-
-	return reqRefProps
-}
-
-func (index *SpecIndex) extractComponentParameters(paramsNode *yaml.Node, pathPrefix string) {
-	var name string
-	for i, param := range paramsNode.Content {
-		if i%2 == 0 {
-			name = param.Value
-			continue
-		}
-		def := fmt.Sprintf("%s%s", pathPrefix, name)
-		ref := &Reference{
-			Definition: def,
-			Name:       name,
-			Node:       param,
-		}
-		index.allParameters[def] = ref
-	}
-}
-
-func (index *SpecIndex) extractComponentRequestBodies(requestBodiesNode *yaml.Node, pathPrefix string) {
-	var name string
-	for i, reqBod := range requestBodiesNode.Content {
-		if i%2 == 0 {
-			name = reqBod.Value
-			continue
-		}
-		def := fmt.Sprintf("%s%s", pathPrefix, name)
-		ref := &Reference{
-			Definition: def,
-			Name:       name,
-			Node:       reqBod,
-		}
-		index.allRequestBodies[def] = ref
-	}
-}
-
-func (index *SpecIndex) extractComponentResponses(responsesNode *yaml.Node, pathPrefix string) {
-	var name string
-	for i, response := range responsesNode.Content {
-		if i%2 == 0 {
-			name = response.Value
-			continue
-		}
-		def := fmt.Sprintf("%s%s", pathPrefix, name)
-		ref := &Reference{
-			Definition: def,
-			Name:       name,
-			Node:       response,
-		}
-		index.allResponses[def] = ref
-	}
-}
-
-func (index *SpecIndex) extractComponentHeaders(headersNode *yaml.Node, pathPrefix string) {
-	var name string
-	for i, header := range headersNode.Content {
-		if i%2 == 0 {
-			name = header.Value
-			continue
-		}
-		def := fmt.Sprintf("%s%s", pathPrefix, name)
-		ref := &Reference{
-			Definition: def,
-			Name:       name,
-			Node:       header,
-		}
-		index.allHeaders[def] = ref
-	}
-}
-
-func (index *SpecIndex) extractComponentCallbacks(callbacksNode *yaml.Node, pathPrefix string) {
-	var name string
-	for i, callback := range callbacksNode.Content {
-		if i%2 == 0 {
-			name = callback.Value
-			continue
-		}
-		def := fmt.Sprintf("%s%s", pathPrefix, name)
-		ref := &Reference{
-			Definition: def,
-			Name:       name,
-			Node:       callback,
-		}
-		index.allCallbacks[def] = ref
-	}
-}
-
-func (index *SpecIndex) extractComponentLinks(linksNode *yaml.Node, pathPrefix string) {
-	var name string
-	for i, link := range linksNode.Content {
-		if i%2 == 0 {
-			name = link.Value
-			continue
-		}
-		def := fmt.Sprintf("%s%s", pathPrefix, name)
-		ref := &Reference{
-			Definition: def,
-			Name:       name,
-			Node:       link,
-		}
-		index.allLinks[def] = ref
-	}
-}
-
-func (index *SpecIndex) extractComponentExamples(examplesNode *yaml.Node, pathPrefix string) {
-	var name string
-	for i, example := range examplesNode.Content {
-		if i%2 == 0 {
-			name = example.Value
-			continue
-		}
-		def := fmt.Sprintf("%s%s", pathPrefix, name)
-		ref := &Reference{
-			Definition: def,
-			Name:       name,
-			Node:       example,
-		}
-		index.allExamples[def] = ref
-	}
-}
-
-func (index *SpecIndex) extractComponentSecuritySchemes(securitySchemesNode *yaml.Node, pathPrefix string) {
-	var name string
-	for i, secScheme := range securitySchemesNode.Content {
-		if i%2 == 0 {
-			name = secScheme.Value
-			continue
-		}
-		def := fmt.Sprintf("%s%s", pathPrefix, name)
-		ref := &Reference{
-			Definition: def,
-			Name:       name,
-			Node:       secScheme,
-			ParentNode: securitySchemesNode,
-			Path:       fmt.Sprintf("$.components.securitySchemes.%s", name),
-		}
-		index.allSecuritySchemes[def] = ref
-	}
-}
-
-func (index *SpecIndex) performExternalLookup(uri []string, componentId string,
-	lookupFunction ExternalLookupFunction, parent *yaml.Node,
-) *Reference {
-	if len(uri) > 0 {
-		externalSpecIndex := index.externalSpecIndex[uri[0]]
-		if externalSpecIndex == nil {
-			_, newRoot, err := lookupFunction(componentId)
-			if err != nil {
-				indexError := &IndexingError{
-					Error: err,
-					Node:  parent,
-					Path:  componentId,
-				}
-				index.refErrors = append(index.refErrors, indexError)
-				return nil
-			}
-
-			// cool, cool, lets index this spec also. This is a recursive action and will keep going
-			// until all remote references have been found.
-			newIndex := NewSpecIndex(newRoot)
-			index.externalSpecIndex[uri[0]] = newIndex
-			externalSpecIndex = newIndex
-		}
-
-		foundRef := externalSpecIndex.FindComponentInRoot(uri[1])
-		if foundRef != nil {
-			nameSegs := strings.Split(uri[1], "/")
-			ref := &Reference{
-				Definition:     componentId,
-				Name:           nameSegs[len(nameSegs)-1],
-				Node:           foundRef.Node,
-				IsRemote:       true,
-				RemoteLocation: componentId,
-				Path:           foundRef.Path,
-			}
-			return ref
-		}
-	}
-	return nil
-}
-
-func (index *SpecIndex) FindComponentInRoot(componentId string) *Reference {
-	if index.root != nil {
-		name, friendlySearch := utils.ConvertComponentIdIntoFriendlyPathSearch(componentId)
-		path, err := yamlpath.NewPath(friendlySearch)
-		if path == nil || err != nil {
-			return nil // no component found
-		}
-		res, _ := path.Find(index.root)
-		if len(res) == 1 {
-			ref := &Reference{
-				Definition:            componentId,
-				Name:                  name,
-				Node:                  res[0],
-				Path:                  friendlySearch,
-				RequiredRefProperties: index.extractDefinitionRequiredRefProperties(res[0], map[string][]string{}),
-			}
-
-			return ref
-		}
-	}
-	return nil
-}
-
-func (index *SpecIndex) countUniqueInlineDuplicates() int {
-	if index.componentsInlineParamUniqueCount > 0 {
-		return index.componentsInlineParamUniqueCount
-	}
-	unique := 0
-	for _, p := range index.paramInlineDuplicates {
-		if len(p) == 1 {
-			unique++
-		}
-	}
-	index.componentsInlineParamUniqueCount = unique
-	return unique
-}
-
-func (index *SpecIndex) scanOperationParams(params []*yaml.Node, pathItemNode *yaml.Node, method string) {
-	for i, param := range params {
-		// param is ref
-		if len(param.Content) > 0 && param.Content[0].Value == "$ref" {
-
-			paramRefName := param.Content[1].Value
-			paramRef := index.allMappedRefs[paramRefName]
-
-			if index.paramOpRefs[pathItemNode.Value] == nil {
-				index.paramOpRefs[pathItemNode.Value] = make(map[string]map[string]*Reference)
-				index.paramOpRefs[pathItemNode.Value][method] = make(map[string]*Reference)
-
-			}
-			// if we know the path, but it's a new method
-			if index.paramOpRefs[pathItemNode.Value][method] == nil {
-				index.paramOpRefs[pathItemNode.Value][method] = make(map[string]*Reference)
-			}
-
-			// if this is a duplicate, add an error and ignore it
-			if index.paramOpRefs[pathItemNode.Value][method][paramRefName] != nil {
-				path := fmt.Sprintf("$.paths.%s.%s.parameters[%d]", pathItemNode.Value, method, i)
-				if method == "top" {
-					path = fmt.Sprintf("$.paths.%s.parameters[%d]", pathItemNode.Value, i)
-				}
-
-				index.operationParamErrors = append(index.operationParamErrors, &IndexingError{
-					Error: fmt.Errorf("the `%s` operation parameter at path `%s`, "+
-						"index %d has a duplicate ref `%s`", method, pathItemNode.Value, i, paramRefName),
-					Node: param,
-					Path: path,
-				})
-			} else {
-				index.paramOpRefs[pathItemNode.Value][method][paramRefName] = paramRef
-			}
-
-			continue
-
-		} else {
-
-			// param is inline.
-			_, vn := utils.FindKeyNode("name", param.Content)
-
-			path := fmt.Sprintf("$.paths.%s.%s.parameters[%d]", pathItemNode.Value, method, i)
-			if method == "top" {
-				path = fmt.Sprintf("$.paths.%s.parameters[%d]", pathItemNode.Value, i)
-			}
-
-			if vn == nil {
-				index.operationParamErrors = append(index.operationParamErrors, &IndexingError{
-					Error: fmt.Errorf("the '%s' operation parameter at path '%s', index %d has no 'name' value",
-						method, pathItemNode.Value, i),
-					Node: param,
-					Path: path,
-				})
-				continue
-			}
-
-			ref := &Reference{
-				Definition: vn.Value,
-				Name:       vn.Value,
-				Node:       param,
-				Path:       path,
-			}
-			if index.paramOpRefs[pathItemNode.Value] == nil {
-				index.paramOpRefs[pathItemNode.Value] = make(map[string]map[string]*Reference)
-				index.paramOpRefs[pathItemNode.Value][method] = make(map[string]*Reference)
-			}
-
-			// if we know the path but this is a new method.
-			if index.paramOpRefs[pathItemNode.Value][method] == nil {
-				index.paramOpRefs[pathItemNode.Value][method] = make(map[string]*Reference)
-			}
-
-			// if this is a duplicate, add an error and ignore it
-			if index.paramOpRefs[pathItemNode.Value][method][ref.Name] != nil {
-				path := fmt.Sprintf("$.paths.%s.%s.parameters[%d]", pathItemNode.Value, method, i)
-				if method == "top" {
-					path = fmt.Sprintf("$.paths.%s.parameters[%d]", pathItemNode.Value, i)
-				}
-
-				index.operationParamErrors = append(index.operationParamErrors, &IndexingError{
-					Error: fmt.Errorf("the `%s` operation parameter at path `%s`, "+
-						"index %d has a duplicate name `%s`", method, pathItemNode.Value, i, vn.Value),
-					Node: param,
-					Path: path,
-				})
-			} else {
-				index.paramOpRefs[pathItemNode.Value][method][ref.Name] = ref
-			}
-			continue
-		}
-	}
-}
-
-func isHttpMethod(val string) bool {
-	switch strings.ToLower(val) {
-	case methodTypes[0]:
-		return true
-	case methodTypes[1]:
-		return true
-	case methodTypes[2]:
-		return true
-	case methodTypes[3]:
-		return true
-	case methodTypes[4]:
-		return true
-	case methodTypes[5]:
-		return true
-	case methodTypes[6]:
-		return true
-	}
-	return false
-}
-
-func (index *SpecIndex) lookupRemoteReference(ref string) (*yaml.Node, *yaml.Node, error) {
-	// split string to remove file reference
-	uri := strings.Split(ref, "#")
-
-	var parsedRemoteDocument *yaml.Node
-	if index.seenRemoteSources[uri[0]] != nil {
-		parsedRemoteDocument = index.seenRemoteSources[uri[0]]
-	} else {
-		resp, err := http.Get(uri[0])
-		if err != nil {
-			return nil, nil, err
-		}
-		body, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			return nil, nil, err
-		}
-
-		var remoteDoc yaml.Node
-		err = yaml.Unmarshal(body, &remoteDoc)
-		if err != nil {
-			return nil, nil, err
-		}
-		parsedRemoteDocument = &remoteDoc
-		index.remoteLock.Lock()
-		index.seenRemoteSources[uri[0]] = &remoteDoc
-		index.remoteLock.Unlock()
-	}
-
-	// lookup item from reference by using a path query.
-	var query string
-	if len(uri) >= 2 {
-		query = fmt.Sprintf("$%s", strings.ReplaceAll(uri[1], "/", "."))
-	} else {
-		query = "$"
-	}
-
-	// remove any URL encoding
-	query = strings.Replace(query, "~1", "./", 1)
-	query = strings.ReplaceAll(query, "~1", "/")
-
-	path, err := yamlpath.NewPath(query)
-	if err != nil {
-		return nil, nil, err
-	}
-	result, _ := path.Find(parsedRemoteDocument)
-	if len(result) == 1 {
-		return result[0], parsedRemoteDocument, nil
-	}
-	return nil, nil, nil
-}
-
-func (index *SpecIndex) lookupFileReference(ref string) (*yaml.Node, *yaml.Node, error) {
-	// split string to remove file reference
-	uri := strings.Split(ref, "#")
-
-	file := strings.ReplaceAll(uri[0], "file:", "")
-
-	var parsedRemoteDocument *yaml.Node
-	if index.seenRemoteSources[file] != nil {
-		parsedRemoteDocument = index.seenRemoteSources[file]
-	} else {
-
-		body, err := ioutil.ReadFile(file)
-		if err != nil {
-			return nil, nil, err
-		}
-
-		var remoteDoc yaml.Node
-		err = yaml.Unmarshal(body, &remoteDoc)
-		if err != nil {
-			return nil, nil, err
-		}
-		parsedRemoteDocument = &remoteDoc
-		index.seenRemoteSources[file] = &remoteDoc
-	}
-
-	// lookup item from reference by using a path query.
-	var query string
-	if len(uri) >= 2 {
-		query = fmt.Sprintf("$%s", strings.ReplaceAll(uri[1], "/", "."))
-	} else {
-		query = "$"
-	}
-
-	// remove any URL encoding
-	query = strings.Replace(query, "~1", "./", 1)
-	query = strings.ReplaceAll(query, "~1", "/")
-
-	path, err := yamlpath.NewPath(query)
-	if err != nil {
-		return nil, nil, err
-	}
-	result, _ := path.Find(parsedRemoteDocument)
-	if len(result) == 1 {
-		return result[0], parsedRemoteDocument, nil
-	}
-
-	return nil, parsedRemoteDocument, nil
 }
